refactor(util): reenforce sanitization
Add prevention for property attack pattern

Ref #4085
